# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/posts": {
      "get": {
        "description": "List social posts with filtering and sorting",
        "operationId": "PostsController_getPosts",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "Filter posts by author id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assetId",
            "required": false,
            "in": "query",
            "description": "Filter posts tagged with an asset id",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "followingOnly",
            "required": true,
            "in": "query",
            "description": "Only return posts from followed authors",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search by content",
            "schema": {
              "maxLength": 255,
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "default": "createdAt",
              "type": "string",
              "enum": [
                "createdAt",
                "viewCount",
                "likeCount",
                "commentCount"
              ]
            }
          },
          {
            "name": "sortDir",
            "required": false,
            "in": "query",
            "schema": {
              "default": "desc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return paginated list of posts"
          },
          "400": {
            "description": "Malformed request payload or failed domain validation."
          },
          "401": {
            "description": "Authentication is missing, invalid, revoked, or expired."
          },
          "403": {
            "description": "Authenticated but not allowed to access or mutate this resource."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Unexpected server error or upstream dependency failure."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List posts",
        "tags": [
          "Posts"
        ]
      },
      "post": {
        "description": "Create a new social post. \n\n**Business Workflow:** Requires user to be fully authenticated. Supports Text, Images, and Video linking.",
        "operationId": "PostsController_createPost",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            },
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            },
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post created successfully"
          },
          "400": {
            "description": "Malformed request payload or failed domain validation."
          },
          "401": {
            "description": "Authentication is missing, invalid, revoked, or expired."
          },
          "403": {
            "description": "Authenticated but not allowed to access or mutate this resource."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Unexpected server error or upstream dependency failure."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Create post",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/me/bookmarks": {
      "get": {
        "operationId": "PostsController_getBookmarkedPosts",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return paginated list of bookmarked posts"
          },
          "400": {
            "description": "Malformed request payload or failed domain validation."
          },
          "401": {
            "description": "Authentication is missing, invalid, revoked, or expired."
          },
          "403": {
            "description": "Authenticated but not allowed to access or mutate this resource."
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Unexpected server error or upstream dependency failure."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List bookmarked posts",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/{postId}": {
      "get": {
        "operationId": "PostsController_getPost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "Post identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return a single post detail"
          },
          "400": {
            "description": "Malformed request payload or failed domain validation."
          },
          "401": {
            "description": "Authentication is missing, invalid, revoked, or expired."
          },
          "403": {
            "description": "Authenticated but not allowed to access or mutate this resource."
          },
          "404": {
            "description": "Requested post was not found."
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Unexpected server error or upstream dependency failure."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get post detail",
        "tags": [
          "Posts"
        ]
      },
      "patch": {
        "operationId": "PostsController_updatePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "Post identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post updated successfully"
          },
          "400": {
            "description": "Malformed request payload or failed domain validation."
          },
          "401": {
            "description": "Authentication is missing, invalid, revoked, or expired."
          },
          "403": {
            "description": "Authenticated but not allowed to access or mutate this resource."
          },
          "404": {
            "description": "Requested post was not found."
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Unexpected server error or upstream dependency failure."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Update post",
        "tags": [
          "Posts"
        ]
      },
      "delete": {
        "operationId": "PostsController_deletePost",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "Post identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Post deleted successfully"
          },
          "400": {
            "description": "Malformed request payload or failed domain validation."
          },
          "401": {
            "description": "Authentication is missing, invalid, revoked, or expired."
          },
          "403": {
            "description": "Authenticated but not allowed to access or mutate this resource."
          },
          "404": {
            "description": "Requested post was not found."
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Unexpected server error or upstream dependency failure."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Delete post",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/{postId}/like": {
      "post": {
        "operationId": "PostsController_togglePostLike",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "Post identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns boolean indicating like status"
          },
          "400": {
            "description": "Malformed request payload or failed domain validation."
          },
          "401": {
            "description": "Authentication is missing, invalid, revoked, or expired."
          },
          "403": {
            "description": "Authenticated but not allowed to access or mutate this resource."
          },
          "404": {
            "description": "Requested post was not found."
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Unexpected server error or upstream dependency failure."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Toggle post like",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/{postId}/bookmark": {
      "post": {
        "operationId": "PostsController_togglePostBookmark",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "Post identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns boolean indicating bookmark status"
          },
          "400": {
            "description": "Malformed request payload or failed domain validation."
          },
          "401": {
            "description": "Authentication is missing, invalid, revoked, or expired."
          },
          "403": {
            "description": "Authenticated but not allowed to access or mutate this resource."
          },
          "404": {
            "description": "Requested post was not found."
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Unexpected server error or upstream dependency failure."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Toggle post bookmark",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/{postId}/comments": {
      "get": {
        "operationId": "PostsController_getComments",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "Post identifier",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "default": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return paginated list of comments"
          },
          "400": {
            "description": "Malformed request payload or failed domain validation."
          },
          "401": {
            "description": "Authentication is missing, invalid, revoked, or expired."
          },
          "403": {
            "description": "Authenticated but not allowed to access or mutate this resource."
          },
          "404": {
            "description": "Requested post was not found."
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Unexpected server error or upstream dependency failure."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "List post comments",
        "tags": [
          "Posts"
        ]
      },
      "post": {
        "operationId": "PostsController_createComment",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "description": "Post identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comment created successfully"
          },
          "400": {
            "description": "Malformed request payload or failed domain validation."
          },
          "401": {
            "description": "Authentication is missing, invalid, revoked, or expired."
          },
          "403": {
            "description": "Authenticated but not allowed to access or mutate this resource."
          },
          "404": {
            "description": "Requested post was not found."
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Unexpected server error or upstream dependency failure."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Create post comment",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/comments/{id}": {
      "patch": {
        "operationId": "PostsController_updateComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Comment identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comment updated successfully"
          },
          "400": {
            "description": "Malformed request payload or failed domain validation."
          },
          "401": {
            "description": "Authentication is missing, invalid, revoked, or expired."
          },
          "403": {
            "description": "Authenticated but not allowed to access or mutate this resource."
          },
          "404": {
            "description": "Requested comment was not found."
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Unexpected server error or upstream dependency failure."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Update comment",
        "tags": [
          "Posts"
        ]
      },
      "delete": {
        "operationId": "PostsController_deleteComment",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Comment identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comment deleted successfully"
          },
          "400": {
            "description": "Malformed request payload or failed domain validation."
          },
          "401": {
            "description": "Authentication is missing, invalid, revoked, or expired."
          },
          "403": {
            "description": "Authenticated but not allowed to access or mutate this resource."
          },
          "404": {
            "description": "Requested comment was not found."
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Unexpected server error or upstream dependency failure."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Delete comment",
        "tags": [
          "Posts"
        ]
      }
    },
    "/posts/comments/{id}/like": {
      "post": {
        "operationId": "PostsController_toggleCommentLike",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Comment identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns boolean indicating like status"
          },
          "400": {
            "description": "Malformed request payload or failed domain validation."
          },
          "401": {
            "description": "Authentication is missing, invalid, revoked, or expired."
          },
          "403": {
            "description": "Authenticated but not allowed to access or mutate this resource."
          },
          "404": {
            "description": "Requested comment was not found."
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "description": "Unexpected server error or upstream dependency failure."
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Toggle comment like",
        "tags": [
          "Posts"
        ]
      }
    },
...
  },
  "info": {
    "title": "RWA Backend API",
    "description": "API documentation for the RWA graduation project - Real World Assets Platform",
    "version": "1.0.0",
    "contact": {
      "name": "RWA Team",
      "url": "",
      "email": "support@rwa.com"
    },
    "license": {
      "name": "Proprietary",
      "url": ""
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Enter JWT token"
      },
      "accessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "name": "apiKey",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreatePostDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10000,
            "description": "Post content"
          },
          "linksUrl": {
            "maxItems": 10,
            "description": "Optional list of images/videos attached to the post",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaLinkDto"
            }
          },
          "assetIds": {
            "maxItems": 20,
            "description": "Optional list of tagged CFD asset identifiers",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "repostId": {
            "type": "string",
            "description": "Optional post identifier to repost"
          }
        },
        "required": [
          "content"
        ]
      },
      "CommentDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "description": "Comment content"
          }
        },
        "required": [
          "content"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "success",
          "code",
          "t"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "code": {
            "type": "string",
            "enum": [
              "internal-error",
              "not-found",
              "bad-request",
              "validation-error",
              "rate-limit-exceeded",
              "duplicate-request",
              "unauthorized",
              "forbidden",
              "token-expired",
              "asset-not-found",
              "invalid-token",
              "token-not-found",
              "session-not-found",
              "session-revoked",
              "session-expired",
              "user-not-found",
              "user-not-active",
              "user-already-exists",
              "invalid-credentials",
              "reference-code-not-found",
              "reference-code-invalid",
              "reference-code-already-set",
              "password-attempt-limit-exceeded",
              "otp-not-found",
              "invalid-otp",
              "invalid-otp-token",
              "mfa-already-enabled",
              "mfa-not-enabled",
              "invalid-mfa-code",
              "challenge-required",
              "challenge-not-found",
              "challenge-expired",
              "challenge-already-used",
              "challenge-invalid-factor",
              "challenge-payload-mismatch",
              "account-not-found",
              "account-not-active",
              "position-not-found",
              "position-already-closed",
              "feedback-not-found",
              "post-not-found",
              "post-unauthorized",
              "comment-not-found",
              "comment-unauthorized",
              "support-ticket-not-found",
              "support-unauthorized",
              "insufficient-balance",
              "transfer-to-same-account",
              "transfer-not-allowed-with-open-position",
              "external-api-error",
              "invalid-quantity",
              "price-not-found"
            ]
          },
          "message": {
            "type": "string",
            "description": "Human-readable error summary for clients/logging."
          },
          "t": {
            "type": "string",
            "format": "date-time"
          },
          "errors": {
            "description": "Detailed error information by field, domain rule, or external provider.",
            "oneOf": [
              {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ValidationErrorItem"
                }
              },
              {
                "type": "object",
                "additionalProperties": true
              }
            ]
          },
          "traceId": {
            "type": "string",
            "description": "Request trace id for debugging across services."
          }
        },
        "additionalProperties": false
      },
...
    },
    "responses": {
...
      "NotFound": {
        "description": "Requested resource was not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "Rate limit exceeded or too many OTP/password attempts.",
        "headers": {
          "Retry-After": {
            "description": "Seconds to wait before retrying.",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  }
}